\name{biomvRCNS-package}
\alias{biomvRCNS-package}
\alias{biomvRCNS}
\docType{package}
\title{
biomvRCNS - Copy Number analysis and Segmentaion models for multivariate Bio data.
}
\description{
In this package, a Hidden Semi Markov Model and one homogeneous segmentation based model are designed and implemeted for genomic data, with the aim of assisting in transcripts detection using high throughput techology like RNA-seq or tiling array, and copy number analyisis using arrary based CGH or targeted sequencing.
}
\details{
\tabular{ll}{
Package: \tab biomvRCNS\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2013-02-03\cr
License: \tab GPL (>= 2)\cr
Depends: \tab methods, mvtnorm, IRanges, GenomicRanges, Gviz\cr
}
}
\author{
Yang Du

Maintainer: Yang Du <duyang@fbn-dummerstorf.de>
}
\references{
	Guedon, Y. (2003). Estimating hidden semi-Markov chains from discrete sequences. Journal of Computational and Graphical Statistics, 12(3), 604-639.
}

\keyword{ package }
\seealso{
  \code{\link{biomvRhsmm}}, 
  \code{\link{biomvRseg}}
}
