\docType{data}
\name{coriell}
\alias{coriell}
\title{Array CGH data set of Coriell cell lines}
\format{A data frame containing five variables: first is clone name,
second is clone chromosome, third is clone position, fourth and fifth are
log2ratio for two cell lines.}
\description{
  These are two data array CGH studies sets of Corriel cell
  lines taken from the reference below.
}
\references{
  \url{http://www.nature.com/ng/journal/v29/n3/suppinfo/ng754\_S1.html}

  Snijders et al., Assembly of microarrays for genome-wide
  measurement of DNA copy number, Nature Genetics, 2001
}
\keyword{data}
\keyword{datasets}

