\docType{data}
\name{variosm}
\alias{variosm}
\title{Differential methylation data from sequencing}
\format{Containing one \code{GRanges} object}
\description{
	Extracted from package \code{BiSeq}, which is a small subset of a published study using targeted bisulfite sequencing data to detect differentially methylated regions (DMRs).
}

\references{
  \url{http://dx.doi.org/10.1182}
Schoofs et al. DNA methylation changes are a late event in acute upromyelocytic leukemia and coincide with loss of transcription factor binding. Blood, Nov 2012.

}
\keyword{data}
\keyword{datasets}
